requirements.txt
requests==requests-2.21.0
websocket-client==0.56.0